added more test cases and increased precision for flow computations

+++ include/LAGraphX.h
@@ -1380,7 +1380,7 @@ int LAGraph_MaxFlow(
     GrB_Index S, //source node index
     GrB_Index T, // sink node index
     //outputs
-    float* f,
+    double* f,
     //inout
     char* msg
 );
